+++ src/client/java/chaos/better_frost_walker/BetterFrostWalkerClient.java
@@ -1,0 +1,10 @@
+package chaos.better_frost_walker;
+
+import net.fabricmc.api.ClientModInitializer;
+
+public class BetterFrostWalkerClient implements ClientModInitializer {
+	@Override
+	public void onInitializeClient() {
+		// This entrypoint is suitable for setting up client-specific logic, such as rendering.
+	}
+}
