--- src/client/java/chaos/frost/NewFrostwalkerClient.java
@@ -1,11 +1,0 @@
-package chaos.frost;
-
-import net.fabricmc.api.ClientModInitializer;
-import net.minecraft.block.TransparentBlock;
-
-public class NewFrostwalkerClient implements ClientModInitializer {
-	@Override
-	public void onInitializeClient() {
-		// This entrypoint is suitable for setting up client-specific logic, such as rendering.
-	}
-}
