--- src/client/java/chaos/frost/mixin/client/ExampleClientMixin.java
@@ -1,15 +1,0 @@
-package chaos.frost.mixin.client;
-
-import net.minecraft.client.MinecraftClient;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-
-@Mixin(MinecraftClient.class)
-public class ExampleClientMixin {
-	@Inject(at = @At("HEAD"), method = "run")
-	private void run(CallbackInfo info) {
-		// This code is injected into the start of MinecraftClient.run()V
-	}
-}
