+++ src/client/resources/better-frost-walker.client.mixins.json
@@ -1,0 +1,11 @@
+{
+	"required": true,
+	"package": "chaos.better_frost_walker.mixin.client",
+	"compatibilityLevel": "JAVA_21",
+	"client": [
+		"ExampleClientMixin"
+	],
+	"injectors": {
+		"defaultRequire": 1
+	}
+}
