--- src/client/resources/frost.client.mixins.json
@@ -1,11 +1,0 @@
-{
-	"required": true,
-	"package": "chaos.frost.mixin.client",
-	"compatibilityLevel": "JAVA_21",
-	"client": [
-		"ExampleClientMixin"
-	],
-	"injectors": {
-		"defaultRequire": 1
-	}
-}
