+++ src/main/java/chaos/better_frost_walker/access/ReplaceDiskEnchantmentEffectAccess.java
@@ -1,0 +1,5 @@
+package chaos.better_frost_walker.access;
+
+public interface ReplaceDiskEnchantmentEffectAccess {
+    void betterfrostwalker$setIsFrostWalker();
+}
