--- src/main/java/chaos/frost/access/ReplaceDiskEnchantmentEffectAccess.java
@@ -1,5 +1,0 @@
-package chaos.frost.access;
-
-public interface ReplaceDiskEnchantmentEffectAccess {
-    void betterfrostwalker$setIsFrostWalker();
-}
