--- src/main/java/chaos/frost/mixin/AlwaysMeltIceMixin.java
@@ -1,18 +1,0 @@
-package chaos.frost.mixin;
-
-import chaos.frost.NewFrostwalker;
-import net.minecraft.block.FrostedIceBlock;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.Constant;
-import org.spongepowered.asm.mixin.injection.ModifyConstant;
-
-@Mixin(FrostedIceBlock.class)
-public abstract class AlwaysMeltIceMixin {
-    @ModifyConstant(method = "scheduledTick", constant = @Constant(intValue = 11))
-    public int mixinLimitInt(int constant) {
-        if (NewFrostwalker.CONFIG.meltIceInTheDark) {
-            return -1;
-        }
-        return constant;
-    }
-}
