+++ src/main/resources/assets/better-frost-walker/blockstates/frosted_magma.json
@@ -1,0 +1,16 @@
+{
+  "variants": {
+    "age=0": {
+      "model": "better-frost-walker:block/frosted_magma_0"
+    },
+    "age=1": {
+      "model": "better-frost-walker:block/frosted_magma_1"
+    },
+    "age=2": {
+      "model": "better-frost-walker:block/frosted_magma_2"
+    },
+    "age=3": {
+      "model": "better-frost-walker:block/frosted_magma_3"
+    }
+  }
+}
