+++ src/main/resources/assets/better-frost-walker/icon.png
