+++ src/main/resources/assets/better-frost-walker/models/block/frosted_magma_3.json
@@ -1,0 +1,6 @@
+{
+  "parent": "minecraft:block/cube_all",
+  "textures": {
+    "all": "better-frost-walker:block/frosted_magma_3"
+  }
+}
