+++ src/main/resources/assets/better-frost-walker/textures/block/frosted_magma_2.png
