--- src/main/resources/assets/frost/blockstates/frosted_magma.json
@@ -1,16 +1,0 @@
-{
-  "variants": {
-    "age=0": {
-      "model": "frost:block/frosted_magma_0"
-    },
-    "age=1": {
-      "model": "frost:block/frosted_magma_1"
-    },
-    "age=2": {
-      "model": "frost:block/frosted_magma_2"
-    },
-    "age=3": {
-      "model": "frost:block/frosted_magma_3"
-    }
-  }
-}
