--- src/main/resources/assets/frost/models/block/frosted_magma_2.json
@@ -1,6 +1,0 @@
-{
-  "parent": "minecraft:block/cube_all",
-  "textures": {
-    "all": "frost:block/magma_2"
-  }
-}
