+++ src/main/resources/data/better-frost-walker/tags/block/replaced_by_frost_walker.json
