--- src/main/resources/frost.mixins.json
@@ -1,15 +1,0 @@
-{
-	"required": true,
-	"package": "chaos.frost.mixin",
-	"compatibilityLevel": "JAVA_21",
-	"mixins": [
-		"AlwaysMeltIceMixin",
-		"EnchantmentMixin",
-		"NoIceFallDamage",
-		"PowderedSnowBlockMixin",
-		"ReplaceDiskEnchantmentEffectMixin"
-	],
-	"injectors": {
-		"defaultRequire": 1
-	}
-}
